Extract common material definitions
